backend/requirements/development.txt
# Include base requirements
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.18.1

# Testing
coverage==7.4.0
pytest-mock==3.12.0

# Code formatting
pre-commit==3.6.0
